# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: School Management System API
  description: |
    Multi-tenant School Management System REST API

    ## Features
    - JWT Authentication
    - Role-based Access Control
    - Multi-tenant Architecture
    - Comprehensive School Management

    ## Roles
    - **Super Admin**: System-level access, can create schools
    - **School Admin**: School-level access, can manage teachers, students, and school operations
    - **Accountant**: Financial management and limited academic operations
    - **Teacher**: Academic operations, attendance, and syllabus tracking
    - **Parent**: Read-only access to child's information
  version: 1.0.0
  contact:
    name: API Support
    email: support@schoolmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.schoolmanagement.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "AUTH_001"
            message: "Unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "AUTH_003"
            message: "Forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "NOT_FOUND"
            message: "Resource not found"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "VALIDATION_ERROR"
            message: "Validation failed"
            errors:
              - field: "email"
                message: "Invalid email format"
...
  schemas:
    Error:
      type: object
      required:
        - success
        - errorCode
        - message
      properties:
        success:
          type: boolean
          example: false
        errorCode:
          type: string
          example: "AUTH_001"
        message:
          type: string
          example: "Invalid email or password"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Success:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
...
    TeacherAllocation:
      type: object
      properties:
        id:
          type: integer
          example: 1
        teacherId:
          type: integer
          example: 5
        teacherName:
          type: string
          example: "Mr. John Smith"
        classId:
          type: integer
          example: 1
        className:
          type: string
          example: "Class 5"
        subjectId:
          type: integer
          example: 2
        subjectName:
          type: string
          example: "Mathematics"
        academicYearId:
          type: integer
          example: 1
        academicYearName:
          type: string
          example: "2024-2025"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    AllocateTeacherRequest:
      type: object
      required:
        - teacherId
        - classId
        - subjectId
        - academicYearId
      properties:
        teacherId:
          type: integer
          example: 5
        classId:
          type: integer
          example: 1
        subjectId:
          type: integer
          example: 2
        academicYearId:
          type: integer
          example: 1

    # Accountant related schemas
...
paths:
...
  /teacher-allocations:
    post:
      tags:
        - Teacher Allocations
      summary: Allocate teacher
      description: Allocate a teacher to a class and subject for a specific academic year (School Admin only). Business Plan feature.
      operationId: allocateTeacher
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocateTeacherRequest"
            examples:
              singleAllocation:
                summary: Allocate teacher to class and subject
                value:
                  teacherId: 5
                  classId: 1
                  subjectId: 2
                  academicYearId: 1
      responses:
        "201":
          description: Teacher allocated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TeacherAllocation"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      tags:
        - Teacher Allocations
      summary: Get all allocations
      description: Retrieve all teacher allocations in the school (Admin, Accountant)
      operationId: getAllAllocations
      security:
        - BearerAuth: []
      parameters:
        - name: academicYearId
          in: query
          description: Filter by academic year
          required: false
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Allocations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TeacherAllocation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /teacher-allocations/teacher/{teacherId}:
    get:
      tags:
        - Teacher Allocations
      summary: Get allocations by teacher
      description: Retrieve all allocations for a specific teacher (Admin, Accountant, Teacher)
      operationId: getAllocationsByTeacher
      security:
        - BearerAuth: []
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
          example: 5
        - name: academicYearId
          in: query
          description: Filter by academic year
          required: false
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Teacher allocations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TeacherAllocation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /teacher-allocations/class/{classId}:
    get:
      tags:
        - Teacher Allocations
      summary: Get allocations by class
      description: Retrieve all teacher allocations for a specific class (Admin, Accountant, Teacher)
      operationId: getAllocationsByClass
      security:
        - BearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: academicYearId
          in: query
          description: Filter by academic year
          required: false
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Class allocations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TeacherAllocation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /teacher-allocations/{id}:
    get:
      tags:
        - Teacher Allocations
      summary: Get allocation by ID
      description: Retrieve teacher allocation details by ID (Admin, Accountant)
      operationId: getAllocationById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Allocation retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TeacherAllocation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Teacher Allocations
      summary: Remove allocation
      description: Remove a teacher allocation (School Admin only)
      operationId: removeAllocation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Teacher allocation removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # Accountants endpoints
...